se realizan los cambios en reportes de factura cuando se compra y vende material

+++ Agregados/CierreApertCajas.cs
@@ -30,7 +30,7 @@ namespace Agregados
         public decimal MontoSinpe { get; set; }
         public decimal MontoCheque { get; set; }
         public decimal MontoCredito { get; set; }
-        public int Accion { get; set; }
+        public byte Accion { get; set; }
         public int IdUsuario { get; set; }
         public Nullable<System.DateTime> FechaSalida { get; set; }
         public string HoraSalida { get; set; }
