+++ Agregados/Forms/Cashiers/FrmCashierClose.cs
@@ -24,7 +24,7 @@ namespace Agregados.Forms.Cashiers
 
         string Fechap;
         string Horap;
-        int Accionp;
+        byte Accionp;
 
         public FrmCashierClose()
         {
